FIX: id should be of unsigned long int

+++ include/toto/my_struct.h
@@ -9,13 +9,14 @@
 // data type for storing read records and write records
 typedef struct{
     int address;
-    int id;
+    unsigned long int id;
 } MyStruct;
 
 // data type for storing region record
 typedef struct{
-    int id;
+    unsigned long int id;
     char function_name[50];
 } FunctionIdSet;
 
 #endif //HOMEWORK1_TIANHUAN_TU_MYSTRUCT_H
+
